handle new temp store

+++ migrations/2025-04-15-172536_create_temp_stores_integrations_table/up.sql
@@ -12,7 +12,7 @@ CREATE TABLE temp_stores_integrations (
     "refresh_token" text NOT NULL,
     "expires" int8 NOT NULL,
     "default_carrier_id" int4 NOT NULL,
-    "integration_platform_id" int4 NOT NULL,
+    "integration_platform_id" int8 NOT NULL,
     "odd_enabled" bool NOT NULL,
     "new_store_id" int8,
     "new_assigned_user_id" int8,
